+++ test/test_other.py
@@ -3162,6 +3162,9 @@ More info: https://emscripten.org
       (['-gsplit-dwarf', '-gsource-map'], True, True, True),
       (['-gsource-map', '-sERROR_ON_WASM_CHANGES_AFTER_LINK'], False, True, True),
       (['-Oz', '-gsource-map'], False, True, True),
+      # Disabling JS debuggability is possible, and we still emit a source map
+      # if requested. (But see above about the name section.)
+      (['-Oz', '-gsource-map', '-sJS_DEBUG_LEVEL=0'], False, True, True),
     ]:
       print(flags, expect_dwarf, expect_sourcemap, expect_names)
       self.emcc(test_file(source_file), flags, js_file)
@@ -9248,6 +9251,7 @@ int main() {
   @crossplatform
   def test_binaryen_debug(self):
     for args, expect_clean_js, expect_whitespace_js, expect_closured in [
+        None,
         (['-O0'], False, True, False),
         (['-O0', '-g1'], False, True, False),
         (['-O0', '-g2'], False, True, False), # in -g2+, we emit -g to asm2wasm so function names are saved
@@ -9256,13 +9260,19 @@ int main() {
         (['-O0', '-gline-tables-only'], False, True, False),
         (['-O1'], False, True, False),
         (['-O3'], True, False, False),
-        (['-Oz', '-gsource-map'], False, True, False), # TODO: fix this (#20462)
+        # source maps by themselves set -g, so the JS is debuggable too
+        (['-Oz', '-gsource-map'], False, True, False),
+        # overriding the JS debug level allows an optimized source map build
+        # (i.e., same as -O2 aside from having a source map)
+        (['-Oz', '-gsource-map', '-sJS_DEBUG_LEVEL=0'], True, False, False),
+        # incrementing the JS debug level to 1 adds whitespace
+        (['-Oz', '-gsource-map', '-sJS_DEBUG_LEVEL=1'], True, True, False),
         (['-O2'], True,  False, False),
         (['-O2', '-gz'], True,  False, False), # -gz means debug compression, it should not enable debugging
         (['-O2', '-g1'], False, True, False),
         (['-O2', '-g'],  False, True, False),
         (['-O2', '--closure=1'], True, False, True),
-        (['-O2', '--closure=1', '-g1'], True, True,  True),
+        (['-O2', '--closure=1', '-g1'], True, True, True),
       ]:
       print(args, expect_clean_js, expect_whitespace_js, expect_closured)
       delete_file('a.out.wat')